+++ src/pretalx/orga/forms/cfp.py
@@ -10,20 +10,16 @@ from pretalx.submission.models import AnswerOption, CfP, Question, SubmissionTyp
 
 class CfPSettingsForm(ReadOnlyFlag, I18nFormMixin, HierarkeyForm):
     cfp_show_deadline = forms.BooleanField(
-        label=_('Display deadline publicly'),
-        required=False,
+        label=_('Display deadline publicly'), required=False
     )
     cfp_request_abstract = forms.BooleanField(
-        label=_('Ask speakers to submit an abstract'),
-        required=False,
+        label=_('Ask speakers to submit an abstract'), required=False
     )
     cfp_request_description = forms.BooleanField(
-        label=_('Ask speakers to submit a longer description'),
-        required=False,
+        label=_('Ask speakers to submit a longer description'), required=False
     )
     cfp_request_notes = forms.BooleanField(
-        label=_('Ask speakers to submit notes for the organiser'),
-        required=False,
+        label=_('Ask speakers to submit notes for the organiser'), required=False
     )
     cfp_request_biography = forms.BooleanField(
         label=_('Ask speakers to provide a biography/personal information'),
@@ -34,131 +30,150 @@ class CfPSettingsForm(ReadOnlyFlag, I18nFormMixin, HierarkeyForm):
         required=False,
     )
     cfp_request_image = forms.BooleanField(
-        label=_('Ask speakers to submit a talk image'),
-        required=False,
+        label=_('Ask speakers to submit a talk image'), required=False
     )
     cfp_require_abstract = forms.BooleanField(
-        label=_('Force speakers to submit an abstract'),
-        required=False,
+        label=_('Force speakers to submit an abstract'), required=False
     )
     cfp_require_description = forms.BooleanField(
-        label=_('Force speakers to submit a longer description'),
-        required=False,
+        label=_('Force speakers to submit a longer description'), required=False
     )
     cfp_require_notes = forms.BooleanField(
-        label=_('Force speakers to submit notes for the organiser'),
-        required=False,
+        label=_('Force speakers to submit notes for the organiser'), required=False
     )
-    cfp_require_biography = forms.BooleanField(
-        required=False,
-    )
+    cfp_require_biography = forms.BooleanField(required=False)
     cfp_require_image = forms.BooleanField(
-        label=_('Force speakers to submit a talk image'),
-        required=False,
+        label=_('Force speakers to submit a talk image'), required=False
     )
-    review_deadline = forms.DateTimeField(
-        label=_('Review deadline'),
-        required=False,
+    cfp_abstract_min_length = forms.IntegerField(
+        label=_('Minimum abstract length in characters'), required=False, min_value=0
     )
+    cfp_description_min_length = forms.IntegerField(
+        label=_('Minimum description length in characters'), required=False, min_value=0
+    )
+    cfp_biography_min_length = forms.IntegerField(
+        label=_('Minimum abstract length in characters'), required=False, min_value=0
+    )
+    cfp_abstract_max_length = forms.IntegerField(
+        label=_('Maximum abstract length in characters'), required=False, min_value=0
+    )
+    cfp_description_max_length = forms.IntegerField(
+        label=_('Maximum description length in characters'), required=False, min_value=0
+    )
+    cfp_biography_max_length = forms.IntegerField(
+        label=_('Maximum abstract length in characters'), required=False, min_value=0
+    )
+    review_deadline = forms.DateTimeField(label=_('Review deadline'), required=False)
     review_score_mandatory = forms.BooleanField(
-        label=_('Require reviewers to submit a score'),
-        required=False,
+        label=_('Require reviewers to submit a score'), required=False
     )
     review_text_mandatory = forms.BooleanField(
-        label=_('Require reviewers to submit a text'),
-        required=False,
+        label=_('Require reviewers to submit a text'), required=False
     )
     mail_on_new_submission = forms.BooleanField(
         label=_('Send mail on new submission'),
-        help_text=_('If this setting is checked, you will receive an email to the organiser address for every received submission.'),
-        required=False
+        help_text=_(
+            'If this setting is checked, you will receive an email to the organiser address for every received submission.'
+        ),
+        required=False,
     )
     allow_override_votes = forms.BooleanField(
         label=_('Allow override votes'),
-        help_text=_('Individual reviewers can be assigned a fixed amount of "override votes" (positive or negative vetos).'),
+        help_text=_(
+            'Individual reviewers can be assigned a fixed amount of "override votes" (positive or negative vetos).'
+        ),
         required=False,
     )
     review_min_score = forms.IntegerField(
-        label=_('Minimum score'),
-        help_text=_('The minimum score reviewers can assign'),
+        label=_('Minimum score'), help_text=_('The minimum score reviewers can assign')
     )
     review_max_score = forms.IntegerField(
-        label=_('Maximum score'),
-        help_text=_('The maximum score reviewers can assign'),
+        label=_('Maximum score'), help_text=_('The maximum score reviewers can assign')
     )
     review_help_text = I18nFormField(
         label=_('Help text for reviewers'),
-        help_text=_('This text will be shown at the top of every review, as long as reviews can be created or edited. You can use markdown here.'),
+        help_text=_(
+            'This text will be shown at the top of every review, as long as reviews can be created or edited. You can use markdown here.'
+        ),
         widget=I18nTextarea,
         required=False,
     )
 
     def __init__(self, obj, *args, **kwargs):
-        kwargs.pop('read_only')  # added in ActionFromUrl view mixin, but not needed here.
+        kwargs.pop(
+            'read_only'
+        )  # added in ActionFromUrl view mixin, but not needed here.
         super().__init__(*args, obj=obj, **kwargs)
         if getattr(obj, 'email'):
-            self.fields['mail_on_new_submission'].help_text += f' (<a href="mailto:{obj.email}">{obj.email}</a>)'
+            self.fields[
+                'mail_on_new_submission'
+            ].help_text += f' (<a href="mailto:{obj.email}">{obj.email}</a>)'
         if getattr(obj, 'slug'):
-            additional = _('You can configure override votes <a href="{link}">in the team settings</a>.').format(link=obj.orga_urls.team_settings)
+            additional = _(
+                'You can configure override votes <a href="{link}">in the team settings</a>.'
+            ).format(link=obj.orga_urls.team_settings)
             self.fields['allow_override_votes'].help_text += f' {additional}'
         minimum = int(obj.settings.review_min_score)
         maximum = int(obj.settings.review_max_score)
-        self.fields['review_deadline'].widget = forms.DateTimeInput(attrs={'class': 'datetimepickerfield'})
+        self.fields['review_deadline'].widget = forms.DateTimeInput(
+            attrs={'class': 'datetimepickerfield'}
+        )
         for number in range(abs(maximum - minimum + 1)):
             index = minimum + number
             self.fields[f'review_score_name_{index}'] = forms.CharField(
                 label=_('Score label ({})').format(index),
-                help_text=_('Human readable explanation of what a score of "{}" actually means, e.g. "great!".').format(index),
-                required=False
+                help_text=_(
+                    'Human readable explanation of what a score of "{}" actually means, e.g. "great!".'
+                ).format(index),
+                required=False,
             )
+        for field in ['abstract', 'description', 'biography']:
+            self.fields[f'cfp_{field}_min_length'].widget.attrs['placeholder'] = ''
+            self.fields[f'cfp_{field}_max_length'].widget.attrs['placeholder'] = ''
 
     def clean(self):
         data = self.cleaned_data
         minimum = int(data.get('review_min_score'))
         maximum = int(data.get('review_max_score'))
         if not minimum < maximum:
-            raise forms.ValidationError(_('Please assign a minimum score smaller than the maximum score!'))
+            raise forms.ValidationError(
+                _('Please assign a minimum score smaller than the maximum score!')
+            )
         return data
 
 
 class CfPForm(ReadOnlyFlag, I18nModelForm):
-
     class Meta:
         model = CfP
-        fields = [
-            'headline', 'text', 'deadline',
-        ]
+        fields = ['headline', 'text', 'deadline']
         widgets = {
             'deadline': forms.DateTimeInput(attrs={'class': 'datetimepickerfield'})
         }
 
 
 class QuestionForm(ReadOnlyFlag, I18nModelForm):
-
     class Meta:
         model = Question
         fields = [
-            'target', 'question', 'help_text', 'variant', 'required',
+            'target',
+            'question',
+            'help_text',
+            'variant',
+            'required',
             'contains_personal_data',
         ]
 
 
 class AnswerOptionForm(ReadOnlyFlag, I18nModelForm):
-
     class Meta:
         model = AnswerOption
-        fields = [
-            'answer',
-        ]
+        fields = ['answer']
 
 
 class SubmissionTypeForm(ReadOnlyFlag, I18nModelForm):
-
     class Meta:
         model = SubmissionType
-        fields = [
-            'name', 'default_duration', 'deadline',
-        ]
+        fields = ['name', 'default_duration', 'deadline']
         widgets = {
             'deadline': forms.DateTimeInput(attrs={'class': 'datetimepickerfield'})
         }