+++ src/pretalx/orga/forms/submission.py
@@ -6,14 +6,17 @@ from pretalx.submission.models import Submission, SubmissionType
 
 
 class SubmissionForm(ReadOnlyFlag, forms.ModelForm):
-
     def __init__(self, event, **kwargs):
         super().__init__(**kwargs)
-        self.fields['submission_type'].queryset = SubmissionType.objects.filter(event=event)
+        self.fields['submission_type'].queryset = SubmissionType.objects.filter(
+            event=event
+        )
 
         if not self.instance.pk:
             self.fields['speaker'] = forms.CharField(
-                help_text=_('Add the email or nickname of the speaker holding the talk. They will be invited to create an account.')
+                help_text=_(
+                    'Add the email or nickname of the speaker holding the talk. They will be invited to create an account.'
+                )
             )
         self.fields['abstract'].widget.attrs['rows'] = 2
         for key in {'abstract', 'description', 'notes', 'image', 'do_not_record'}:
@@ -23,11 +26,27 @@ class SubmissionForm(ReadOnlyFlag, forms.ModelForm):
                 self.fields.pop(key)
             else:
                 self.fields[key].required = require
+                for attr in ['min', 'max']:
+                    value = event.settings.get(f'cfp_{key}_{attr}_length')
+                    if value:
+                        self.fields[key].widget.attrs[f'{attr}length'] = value
+                        self.fields[key].help_text += ' ' + (
+                            _('Please write at least {} characters.')
+                            if attr == 'min'
+                            else _('Please write no more than {} characters.')
+                        ).format(value)
 
     class Meta:
         model = Submission
         fields = [
-            'title', 'submission_type', 'abstract', 'description',
-            'notes', 'content_locale', 'do_not_record', 'duration',
-            'image', 'is_featured',
+            'title',
+            'submission_type',
+            'abstract',
+            'description',
+            'notes',
+            'content_locale',
+            'do_not_record',
+            'duration',
+            'image',
+            'is_featured',
         ]